requirements.txt
Django==1.9.4
djangorestframework==3.3.2
httpie==0.9.3
mysqlclient==1.3.7
Pygments==2.1.3
requests==2.9.1
